refactor: user/admin/roomescape 클래스 별로 패키지 분리
