+++ src/main/java/roomescape/dto/response/ReservationTimeResponse.java
@@ -1,9 +1,8 @@
 package roomescape.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import roomescape.domain.ReservationTime;
-
 import java.time.LocalTime;
+import roomescape.domain.roomescape.ReservationTime;
 
 public record ReservationTimeResponse(
         long timeId,
