+++ src/main/java/roomescape/service/roomescape/ThemeService.java
@@ -1,15 +1,15 @@
-package roomescape.service;
+package roomescape.service.roomescape;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
-import roomescape.domain.Theme;
+import roomescape.domain.roomescape.Theme;
 import roomescape.dto.request.ThemeSaveRequest;
 import roomescape.dto.response.ThemeDeleteResponse;
 import roomescape.dto.response.ThemeResponse;
-import roomescape.repository.ReservationDao;
-import roomescape.repository.ThemeDao;
+import roomescape.repository.roomescape.ReservationDao;
+import roomescape.repository.roomescape.ThemeDao;
 
 @Service
 public class ThemeService {
