--- src/test/java/roomescape/controller/WelcomePageTest.java
@@ -1,19 +1,0 @@
-package roomescape.controller;
-
-import io.restassured.RestAssured;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class WelcomePageTest {
-    @Test
-    void getWelcomePageTest() {
-        // when & then
-        RestAssured.given().log().all()
-                .when().get("/")
-                .then().log().all()
-                .statusCode(200);
-    }
-}
